+++ AvgScript/src/lib/regExp.ts
@@ -1,3 +1,3 @@
-export const regexNumber = /\+[0-9]+(.[0-9]+)?|-[0-9]+(.[0-9]+)?|[0-9]+(.[0-9]+)?/gi;
+export const regexNumber = /(\+|\-)?([0-9]+)(\.[0-9]+)?/gi;
 export const regexHexColor = /(#|0[x|X])[0-9a-fA-F]{6}/gi;
 export const regexRep = /\<.*\>/gi;